Revert "Revert "[core] Deprecate LIFO/FIFO worker killing policies" (#56960)"
This reverts commit b34929b7a0c9e41e1920bd61d37e3834614b0434.
