+++ release/release_data_tests.yaml
@@ -413,8 +413,6 @@
 
   cluster:
     byod:
-      runtime_env:
-        - RAY_worker_killing_policy=retriable_lifo
       pip:
         - ray[default]
     cluster_compute: "{{scaling}}_all_to_all_compute.yaml"
@@ -431,8 +429,6 @@
 
   cluster:
     byod:
-      runtime_env:
-        - RAY_worker_killing_policy=retriable_lifo
       pip:
         - ray[default]
     cluster_compute: dataset/autoscaling_all_to_all_compute.yaml
@@ -455,8 +451,6 @@
 
   cluster:
     byod:
-      runtime_env:
-        - RAY_worker_killing_policy=retriable_lifo
       pip:
         - ray[default]
     cluster_compute: "{{scaling}}_all_to_all_compute.yaml"
@@ -472,8 +466,6 @@
 
   cluster:
     byod:
-      runtime_env:
-        - RAY_worker_killing_policy=retriable_lifo
       pip:
         - ray[default]
     cluster_compute: dataset/autoscaling_all_to_all_compute.yaml
