--- src/ray/raylet/worker_killing_policy_retriable_fifo.h
@@ -1,44 +1,0 @@
-// Copyright 2022 The Ray Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include <gtest/gtest_prod.h>
-
-#include <utility>
-#include <vector>
-
-#include "absl/container/flat_hash_set.h"
-#include "absl/time/clock.h"
-#include "absl/time/time.h"
-#include "ray/common/memory_monitor.h"
-#include "ray/raylet/worker.h"
-#include "ray/raylet/worker_killing_policy.h"
-
-namespace ray {
-
-namespace raylet {
-
-/// Prefers killing retriable workers over non-retriable ones, then in FIFO order.
-class RetriableFIFOWorkerKillingPolicy : public WorkerKillingPolicy {
- public:
-  RetriableFIFOWorkerKillingPolicy();
-  const std::pair<std::shared_ptr<WorkerInterface>, bool> SelectWorkerToKill(
-      const std::vector<std::shared_ptr<WorkerInterface>> &workers,
-      const MemorySnapshot &system_memory) const;
-};
-
-}  // namespace raylet
-
-}  // namespace ray
